Updated employee management code and schemas
The code related to the management of employees has been updated. Changes have been made to the EmployeesModule and EmployeeService, and the schemas have been modified to handle the relations between companies and employees more accurately. These updates facilitate better data management and increase the flexibility when working with company-employee relations.